Data Engineering/KHS/DE_local/local/my-riot-pipeline/airflow/requirements.txt
pandas
tqdm
kafka-python
